fyp_api/requirements.txt
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.0
colorama==0.4.4
distlib==0.3.4
fastapi==0.75.0
filelock==3.6.0
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
imageio==2.19.0
importlib-metadata==4.11.3
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.4
opencv-python==4.5.5.64
passlib==1.7.4
pathspec==0.9.0
Pillow==9.1.0
pipenv==2022.3.28
platformdirs==2.5.1
praw==7.5.0
prawcore==2.3.0
priority==2.0.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyflakes==2.4.0
PyJWT==2.3.0
pypika-tortoise==0.1.4
python-dotenv==0.20.0
python-multipart==0.0.5
python-pip==1.1.1
pytz==2022.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.34
starlette==0.17.1
toml==0.10.2
tomli==2.0.1
tortoise-orm==0.19.0
typing_extensions==4.1.1
update-checker==0.18.0
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.14.0
virtualenv-clone==0.5.7
websocket-client==1.3.2
Werkzeug==2.1.1
wsproto==1.1.0
zipp==3.8.0
gunicorn==20.1.0